backend-langgraph/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
aiohttp>=3.9.0
pyyaml>=6.0.0
python-json-logger>=2.0.7

